pyproject.toml
[project]
name = "tg-bot-parking-pay"
version = "0.1.0"
description = ""
authors = [
    {name = "george2066"}
]
readme = "README.md"
requires-python = ">=3.9,<=3.13"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "qreader (>=3.16,<4.0)",
    "aiogram (>=3.19.0,<4.0.0)",
    "pyzbar (>=0.1.9,<0.2.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
